Refactor product package names
