+++ geode-connectors/src/main/java/org/apache/geode/connectors/jdbc/internal/cli/DescribeMappingResult.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.connectors.util.internal;
+package org.apache.geode.connectors.jdbc.internal.cli;
 
 import java.util.LinkedHashMap;
 import java.util.List;
