+++ geode-connectors/src/main/resources/META-INF/services/org.apache.geode.distributed.internal.DistributedSystemService
@@ -12,5 +12,5 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-org.apache.geode.internal.ConnectorsDistributedSystemService
+org.apache.geode.connectors.jdbc.internal.ConnectorsDistributedSystemService
